Extract LibDDWAF and Handle
